Rename hooks to `together_*` and support config
Config can be customized with a custom TogetherCLI subclass (to change
the default) or by just using the `together_configure` hook to modify
the config (which defaults to a dict).

Config is then re-exposed during command registration, so that
commands can be built such that they vary based on the config.